Fixed various things, added test batch script
Message reading was broken, and sounded terrible once it was working. I changed it from reading 1 word at a time to instead reading 1 sentence at a time. God help you if you get a donation without punctuation.

Also made a quick batch script to test if it's working. You need to add your access key to it for it to work.